x86/amd_nb: Move SMN access code to a new amd_node driver
commit d6caeafaa324e6aba5ed2ca1a416340c2fd061a2 upstream.

SMN access was bolted into amd_nb mostly as convenience.  This has
limitations though that require incurring tech debt to keep it working.

Move SMN access to the newly introduced AMD Node driver.

[Backport changes]

In drivers/hwmon/k10temp.c, the function read_tempreg_nb_zen uses
amd_pci_dev_to_node_id, which causes an implicit declaration error.
The header move from <asm/amd_node.h> to <asm/amd/node.h> introduced by
commit 0a35c9280a9105 only relocates the file without functional changes,
so we’re skipping this update and keeping the old header path.

Acked-by: Ilpo Järvinen <ilpo.jarvinen@linux.intel.com> # pdx86
Acked-by: Shyam Sundar S K <Shyam-sundar.S-k@amd.com> # PMF, PMC
Link: https://lore.kernel.org/r/20241206161210.163701-11-yazen.ghannam@amd.com
